+++ coreservice/core_service.go
@@ -18,80 +18,21 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/rokwire/core-auth-library-go/authservice"
+	"github.com/rokwire/core-auth-library-go/authutils"
 	"github.com/rokwire/logging-library-go/logs"
 )
 
 // CoreService contains configurations and helper functions required to utilize certain core services
 type CoreService struct {
-	serviceAccountManager authservice.ServiceAccountManager
+	serviceAccountManager *authservice.ServiceAccountManager
 
 	deletedAccountsConfig *DeletedAccountsConfig
 
 	logger *logs.Logger
-}
-
-func (c *CoreService) getDeletedAccountsWithRetry() ([]string, error) {
-	accountIDs, err := c.requestDeletedAccounts()
-	if err != nil {
-		if strings.HasPrefix(err.Error(), "error getting deleted accounts: 401") {
-			// access token may have expired, so get a new one and try once more
-			_, tokenErr := c.serviceAccountManager.GetAccessToken()
-			if tokenErr != nil {
-				return nil, fmt.Errorf("error getting new access token - %v - after %v", tokenErr, err)
-			}
-
-			accountIDs, err = c.requestDeletedAccounts()
-			if err != nil {
-				return nil, err
-			}
-
-			return accountIDs, nil
-		}
-
-		return nil, err
-	}
-
-	return accountIDs, nil
-}
-
-func (c *CoreService) requestDeletedAccounts() ([]string, error) {
-	client := &http.Client{}
-	req, err := http.NewRequest("GET", c.deletedAccountsConfig.Host+c.deletedAccountsConfig.Path, nil)
-	if err != nil {
-		return nil, fmt.Errorf("error formatting request to get deleted accounts: %v", err)
-	}
-
-	req.Header.Set("Authorization", c.serviceAccountManager.CachedAccessToken().String())
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("error requesting deleted accounts: %v", err)
-	}
-
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("error reading body of deleted accounts response: %v", err)
-	}
-
-	if resp.StatusCode != 200 {
-		return nil, fmt.Errorf("error getting deleted accounts: %d - %s", resp.StatusCode, string(body))
-	}
-
-	var deletedAccounts []string
-	err = json.Unmarshal(body, &deletedAccounts)
-	if err != nil {
-		return nil, fmt.Errorf("error on unmarshal deleted accounts response: %v", err)
-	}
-
-	return deletedAccounts, nil
 }
 
 // StartDeletedAccountsTimer starts a timer that periodically retrieves deleted account IDs
@@ -102,18 +43,18 @@ func (c *CoreService) StartDeletedAccountsTimer() {
 		c.deletedAccountsConfig.timer.Stop()
 	}
 
-	c.getDeletedAccounts(c.deletedAccountsConfig.Callback)
+	c.getDeletedAccountsWithCallback(c.deletedAccountsConfig.Callback)
 }
 
-func (c *CoreService) getDeletedAccounts(callback func([]string) error) {
-	accountIDs, err := c.getDeletedAccountsWithRetry()
+func (c *CoreService) getDeletedAccountsWithCallback(callback func([]string) error) {
+	accountIDs, err := c.getDeletedAccounts()
 	if err != nil && c.logger != nil {
 		c.logger.Error(err.Error())
 	}
 
 	err = callback(accountIDs)
 	if err != nil && c.logger != nil {
-		c.logger.Errorf("Received error from callback function: %v", err)
+		c.logger.Errorf("received error from callback function: %v", err)
 	}
 
 	duration := time.Hour * time.Duration(int64(c.deletedAccountsConfig.Period))
@@ -123,24 +64,72 @@ func (c *CoreService) getDeletedAccounts(callback func([]string) error) {
 		// timer expired
 		c.deletedAccountsConfig.timer = nil
 
-		c.getDeletedAccounts(callback)
+		c.getDeletedAccountsWithCallback(callback)
 	case <-c.deletedAccountsConfig.timerDone:
 		// timer aborted
 		c.deletedAccountsConfig.timer = nil
 	}
 }
 
+func (c *CoreService) getDeletedAccounts() ([]string, error) {
+	accountIDs := make([]string, 0)
+
+	req, err := c.buildDeletedAccountsRequest()
+	if err != nil {
+		return nil, fmt.Errorf("error building deleted accounts request: %v", err)
+	}
+
+	responses := c.serviceAccountManager.MakeRequests(req, nil)
+	for _, reqResp := range responses {
+		if reqResp.Error != nil && c.logger != nil {
+			c.logger.Errorf("error making deleted accounts request: %v", reqResp.Error)
+			continue
+		}
+
+		body, err := authutils.ReadResponseBody(reqResp.Response)
+		if err != nil {
+			return nil, fmt.Errorf("error reading deleted accounts response body: %v", err)
+		}
+
+		var deleted []string
+		err = json.Unmarshal(body, &deleted)
+		if err != nil {
+			return nil, fmt.Errorf("error unmarshaling deleted accounts response body: %v", err)
+		}
+
+		accountIDs = append(accountIDs, deleted...)
+	}
+
+	return accountIDs, nil
+}
+
+func (c *CoreService) buildDeletedAccountsRequest() (*http.Request, error) {
+	req, err := http.NewRequest("GET", c.serviceAccountManager.AuthService.AuthBaseURL+c.deletedAccountsConfig.path, nil)
+	if err != nil {
+		return nil, fmt.Errorf("error formatting request to get deleted accounts: %v", err)
+	}
+
+	return req, nil
+}
+
 // NewCoreService creates and configures a new CoreService instance
-func NewCoreService(serviceAccountManager authservice.ServiceAccountManager, deletedAccountsConfig *DeletedAccountsConfig, firstParty bool, logger *logs.Logger) (*CoreService, error) {
+func NewCoreService(serviceAccountManager *authservice.ServiceAccountManager, deletedAccountsConfig *DeletedAccountsConfig, logger *logs.Logger) (*CoreService, error) {
 	if serviceAccountManager == nil {
 		return nil, errors.New("service account manager is missing")
 	}
 
 	if deletedAccountsConfig != nil {
-		if deletedAccountsConfig.Host == "" {
-			return nil, errors.New("deleted accounts host is missing")
+		deletedAccountsConfig.path = "/tps/deleted-accounts"
+		if serviceAccountManager.AuthService.FirstParty {
+			deletedAccountsConfig.path = "/bbs/deleted-accounts"
 		}
-		checkDeletedAccountsConfig(deletedAccountsConfig, firstParty)
+
+		if deletedAccountsConfig.Callback != nil {
+			deletedAccountsConfig.timerDone = make(chan bool)
+			if deletedAccountsConfig.Period == 0 {
+				deletedAccountsConfig.Period = 2
+			}
+		}
 	}
 
 	core := CoreService{serviceAccountManager: serviceAccountManager, deletedAccountsConfig: deletedAccountsConfig, logger: logger}
@@ -148,30 +137,12 @@ func NewCoreService(serviceAccountManager authservice.ServiceAccountManager, del
 	return &core, nil
 }
 
-func checkDeletedAccountsConfig(config *DeletedAccountsConfig, firstParty bool) {
-	if config.Path == "" {
-		if firstParty {
-			config.Path = "/bbs/deleted-accounts"
-		} else {
-			config.Path = "/tps/deleted-accounts"
-		}
-	}
-
-	if config.Callback != nil {
-		config.timerDone = make(chan bool)
-		if config.Period == 0 {
-			config.Period = 2
-		}
-	}
-}
-
-//DeletedAccountsConfig represents a configuration for getting deleted accounts from a remote host
+//DeletedAccountsConfig represents a configuration for getting deleted accounts from a remote core service
 type DeletedAccountsConfig struct {
-	Host     string               //URL of accounts host
-	Path     string               // Path to deleted accounts API
 	Callback func([]string) error // Function to call once the deleted accounts are received
-	Period   uint                 // How often to request deleted account list (in hours)
+	Period   uint                 // How often to request deleted account list in hours (the default is 2)
 
+	path      string
 	timerDone chan bool
 	timer     *time.Timer
 }