+++ sigauth/signature.go
@@ -20,7 +20,6 @@ import (
 	"crypto/rand"
 	"crypto/rsa"
 	"encoding/base64"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
@@ -121,7 +120,7 @@ func (s *SignatureAuth) SignRequest(r *http.Request) error {
 	if err != nil {
 		return fmt.Errorf("error getting service key fingerprint: %v", err)
 	}
-	sigAuthHeader := SignatureAuthHeader{KeyId: serviceKeyFingerprint, Algorithm: "rsa-sha256", Headers: headers}
+	sigAuthHeader := SignatureAuthHeader{KeyID: serviceKeyFingerprint, Algorithm: "rsa-sha256", Headers: headers}
 
 	sigString, err := BuildSignatureString(signedRequest, headers)
 	if err != nil {
@@ -177,14 +176,14 @@ func (s *SignatureAuth) CheckRequestServiceSignature(r *Request, requiredService
 			return "", fmt.Errorf("error getting service key fingerprint: %v", err)
 		}
 
-		if fingerprint == sigAuthHeader.KeyId {
+		if fingerprint == sigAuthHeader.KeyID {
 			found = true
 			break
 		}
 	}
 
 	if !found {
-		return "", fmt.Errorf("request signer fingerprint (%s) does not match any of the required services %v", sigAuthHeader.KeyId, requiredServiceIDs)
+		return "", fmt.Errorf("request signer fingerprint (%s) does not match any of the required services %v", sigAuthHeader.KeyID, requiredServiceIDs)
 	}
 
 	err = s.CheckSignature(serviceReg.PubKey.Key, []byte(sigString), sigAuthHeader.Signature)
@@ -253,40 +252,22 @@ func (s *SignatureAuth) checkRequest(r *Request) (string, *SignatureAuthHeader, 
 	return sigString, sigAuthHeader, nil
 }
 
-// BuildAccessTokenRequest builds a signed request to get an access token from an auth service
-func (s *SignatureAuth) BuildAccessTokenRequest(host string, path string, accountID string, _ string) (*http.Request, error) {
-	if host == "" {
-		return nil, errors.New("host is missing")
-	}
-	if path == "" {
-		return nil, errors.New("path is missing")
-	}
-	if accountID == "" {
-		return nil, errors.New("service account ID is missing")
-	}
+// Implement ServiceAuthRequests interface
 
-	params := map[string]interface{}{
-		"account_id": accountID,
-		"auth_type":  "signature",
+// BuildRequestAuthBody returns a map containing the auth fields for static token auth request bodies
+func (s *SignatureAuth) BuildRequestAuthBody() map[string]interface{} {
+	return map[string]interface{}{
+		"auth_type": "signature",
 	}
-	data, err := json.Marshal(params)
+}
+
+// ModifyRequest signs the passed request to perform signature auth
+func (s *SignatureAuth) ModifyRequest(req *http.Request) error {
+	err := s.SignRequest(req)
 	if err != nil {
-		return nil, fmt.Errorf("error marshaling body for get access token: %v", err)
+		return fmt.Errorf("error signing request: %v", err)
 	}
-
-	r, err := http.NewRequest(http.MethodPost, host+path, bytes.NewReader(data))
-	if err != nil {
-		return nil, fmt.Errorf("error creating request for get access token: %v", err)
-	}
-
-	r.Header.Set("Content-Type", "application/json")
-
-	err = s.SignRequest(r)
-	if err != nil {
-		return nil, fmt.Errorf("error signing request for get access token: %v", err)
-	}
-
-	return r, nil
+	return nil
 }
 
 // NewSignatureAuth creates and configures a new SignatureAuth instance
@@ -398,7 +379,7 @@ func ParseHTTPRequest(r *http.Request) (*Request, error) {
 
 //SignatureAuthHeader defines the structure of the Authorization header for signature authentication
 type SignatureAuthHeader struct {
-	KeyId      string   `json:"keyId" validate:"required"`
+	KeyID      string   `json:"keyId" validate:"required"`
 	Algorithm  string   `json:"algorithm" validate:"required"`
 	Headers    []string `json:"headers,omitempty"`
 	Extensions string   `json:"extensions,omitempty"`
@@ -409,7 +390,7 @@ type SignatureAuthHeader struct {
 func (s *SignatureAuthHeader) SetField(field string, value string) error {
 	switch field {
 	case "keyId":
-		s.KeyId = value
+		s.KeyID = value
 	case "algorithm":
 		s.Algorithm = value
 	case "headers":
@@ -443,7 +424,7 @@ func (s *SignatureAuthHeader) Build() (string, error) {
 		extensions = fmt.Sprintf("extensions=\"%s\",", s.Extensions)
 	}
 
-	return fmt.Sprintf("Signature keyId=\"%s\",algorithm=\"%s\",%s%ssignature=\"%s\"", s.KeyId, s.Algorithm, headers, extensions, s.Signature), nil
+	return fmt.Sprintf("Signature keyId=\"%s\",algorithm=\"%s\",%s%ssignature=\"%s\"", s.KeyID, s.Algorithm, headers, extensions, s.Signature), nil
 }
 
 // ParseSignatureAuthHeader parses a signature Authorization header string